+++ semgrep-core/src/api/AST_generic_to_v1.ml
@@ -631,7 +631,10 @@ and map_keyword_attribute = function
   | DefaultImpl -> Left `DefaultImpl
   (* new: *)
   | Lazy -> Right "lazy"
-  | CaseClass -> Right "CaseClass"
+  | RecordClass -> Right "RecordClass"
+  | AnnotationClass -> Right "AnnotationClass"
+  | EnumClass -> Right "EnumClass"
+  | SealedClass -> Right "SealedClass"
 
 and map_other_attribute_operator _x = "TODO"
 
@@ -1108,7 +1111,7 @@ and map_class_definition
   let v_cparams = map_parameters v_cparams in
   let v_cmixins = map_of_list map_type_ v_cmixins in
   let v_cimplements = map_of_list map_type_ v_cimplements in
-  let v_cextends = map_of_list map_type_ v_cextends in
+  let v_cextends = map_of_list map_class_parent v_cextends in
   let v_ckind = map_wrap map_class_kind v_ckind in
   {
     B.ckind = v_ckind;
@@ -1119,14 +1122,13 @@ and map_class_definition
     cparams = v_cparams;
   }
 
+and map_class_parent (_v1, _v2) = failwith "TODO"
+
 and map_class_kind = function
   | Class -> `Class
   | Interface -> `Interface
   | Trait -> `Trait
   | Object -> `Object
-  | AtInterface -> `AtInterface
-  | RecordClass -> `RecordClass
-  | EnumClass -> failwith "TODO"
 
 and map_directive { d; d_attrs } =
   let d = map_directive_kind d in